+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - [core] feature: provide better warning messages when failed to initialize a module on server startup.
 - [core] feature: use a better implementation to invoke module initializers, and skip the initialization of variables if the module is disabled.
+- [fuji] fix: should escape the `placeholders` and `style tags` when displaying `document string`.